+++ biochemistry-problems/alpha_helix_h-bonds.py
@@ -101,12 +101,12 @@ def write_question(N, args):
 #=====================
 def parse_arguments():
 	parser = bptools.make_arg_parser()
-parser = bptools.add_choice_args(parser, default=4)
+	parser = bptools.add_choice_args(parser, default=4)
 	parser = bptools.add_question_format_args(
 		parser,
 		types_list=['ma', 'mc'],
-		dest='question_type',
-		required=True
+		required=False,
+		default='mc'
 	)
 
 	args = parser.parse_args()